Moved Events managers into their own sub package
