+++ jailbreak-service/src/main/java/org/jailbreak/service/ServiceModule.java
@@ -19,26 +19,26 @@ import org.jailbreak.client.base.FacebookClientImpl;
 import org.jailbreak.service.auth.ApiTokenAuthenticator;
 import org.jailbreak.service.base.ApiTokensManagerImpl;
 import org.jailbreak.service.base.CheckinsManagerImpl;
-import org.jailbreak.service.base.DonateEventsManagerImpl;
 import org.jailbreak.service.base.DonationsManagerImpl;
-import org.jailbreak.service.base.EventsManagerImpl;
-import org.jailbreak.service.base.LinkEventsManagerImpl;
 import org.jailbreak.service.base.SecureTokenGeneratorImpl;
 import org.jailbreak.service.base.StripeManagerImpl;
 import org.jailbreak.service.base.TeamsManagerImpl;
 import org.jailbreak.service.base.UsersManagerImpl;
-import org.jailbreak.service.base.YoutubeEventsManagerImpl;
+import org.jailbreak.service.base.events.DonateEventsManagerImpl;
+import org.jailbreak.service.base.events.EventsManagerImpl;
+import org.jailbreak.service.base.events.LinkEventsManagerImpl;
+import org.jailbreak.service.base.events.YoutubeEventsManagerImpl;
 import org.jailbreak.service.core.ApiTokensManager;
 import org.jailbreak.service.core.CheckinsManager;
-import org.jailbreak.service.core.DonateEventsManager;
 import org.jailbreak.service.core.DonationsManager;
-import org.jailbreak.service.core.EventsManager;
-import org.jailbreak.service.core.LinkEventsManager;
 import org.jailbreak.service.core.SecureTokenGenerator;
 import org.jailbreak.service.core.StripeManager;
 import org.jailbreak.service.core.TeamsManager;
 import org.jailbreak.service.core.UsersManager;
-import org.jailbreak.service.core.YoutubeEventsManager;
+import org.jailbreak.service.core.events.DonateEventsManager;
+import org.jailbreak.service.core.events.EventsManager;
+import org.jailbreak.service.core.events.LinkEventsManager;
+import org.jailbreak.service.core.events.YoutubeEventsManager;
 import org.jailbreak.service.db.dao.ApiTokensDAO;
 import org.jailbreak.service.db.dao.CheckinsDAO;
 import org.jailbreak.service.db.dao.DonationsDAO;
