+++ jailbreak-service/src/main/java/org/jailbreak/service/core/events/LinkEventsManager.java
@@ -1,4 +1,4 @@
-package org.jailbreak.service.core;
+package org.jailbreak.service.core.events;
 
 import java.util.HashMap;
 import java.util.Set;
